copy memory area

+++ 0x07-pointers_arrays_strings/main.h
@@ -10,8 +10,8 @@
 #include <stddef.h>
 
 int _putchar(char c);
-char *_memset(char *s, char b, size_t n);
-char *_memcpy(char *dest, char  *src, unsigned int n);
+char *_memset(char *s, char b, unsigned int n);
+char *_memcpy(char *dest, char *src, unsigned int n);
 char *_strchr(char *s, char c);
 unsigned int _strspn(char *s, char *accept);
 char *_strpbrk(char *s, char *accept);
